pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "generate-thumbnail"
version = "0.1.0"
description = "为当前目录下的所有新增文件创建缩略图"
authors = [{ name = "hanasaki", email = "hanasakayui2022@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "pathlib>=1.0.1",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "watchdog>=5.0.2",
]

[project.scripts]
start-watching = "generate_thumbnail.__init__:start_watching"
